backend/requirements.txt
django == 5.1.6
psycopg2 == 2.9.10
python-dotenv == 1.0.1
django-localflavor == 4.0
pillow == 11.1.0
djangorestframework==3.15.2
django-rest-swagger==2.2.0
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.9